refactor(dashboard): split SessionTable into focused modules — 1115 → 822 LOC
Extract SessionMobileCard (114 lines) and shared utilities
(sessionTableUtils, 110 lines) from SessionTable.tsx.

Module breakdown:
- SessionTable.tsx: 822 lines (orchestrator: data fetching, filtering,
  pagination, keyboard nav, bulk actions, demo sessions)
- SessionMobileCard.tsx: 114 lines (mobile card view, memoized)
- sessionTableUtils.ts: 110 lines (types, formatting, search, comparison)

Changes:
- SessionMobileCard extracted with its own imports and memo boundary
- SessionDesktopRow removed (dead code — not used since VirtualizedSessionList)
- Shared types (SessionRowProps, SessionsPaginationState, SessionRowViewModel)
  and utility functions (formatStatusLabel, matchesSearch, etc.) in utils
- Fixed formatStatusLabel to return 'Permission Prompt' (was 'Permission')
- Fixed isDisplayedSessionEqual to use lastActivity (updatedAt doesn't exist)

Bundle: SessionTable 47.25 kB → 43.86 kB (-7%)

Closes #2932
